Project II  Part II/requirements.txt
opencv-python
imutils
facenet-pytorch
torch
boto3
Pillow
numpy